# Always coerce id fields when performing mongotize even if query_objectid_as_string is on.