--- a/eve/tests/methods/get.py
+++ b/eve/tests/methods/get.py
@@ -240,6 +240,19 @@ class TestGet(TestBase):
         resource = response["_items"]
         self.assertEqual(len(resource), 0)
 
+    def test_get_where_mongo_objectid_as_string_but_field_is_id(self):
+        where_in = '{"tid": { "$in": ["%s"]} }' % self.item_tid
+        response, status = self.get(self.known_resource, "?where=%s" % where_in)
+        self.assert200(status)
+        resource = response["_items"]
+        self.assertEqual(len(resource), 1)
+
+        self.app.config["DOMAIN"]["contacts"]["query_objectid_as_string"] = True
+        response, status = self.get(self.known_resource, "?where=%s" % where_in)
+        self.assert200(status)
+        resource = response["_items"]
+        self.assertEqual(len(resource), 0)
+
     def test_get_where_python_syntax(self):
         where = "ref == %s" % self.item_name
         response, status = self.get(self.known_resource, "?where=%s" % where)