@ehiggs 2020-01-16
I understand performance in this area is important, so here are some microbenchmarks to try this out. 

Given the following setup:

```python
DOMAIN = <copy pasted from debug session when running unit tests>
import eve.io.mongo.mongo as eimm
eimm.config.DOMAIN = DOMAIN
eimm.versioned_id_field = lambda x: 'id_field_versioned'
resource = 'contacts'                                                                                                                                
source = {'_id': '5e20d925927678c6959b3d36', 'rows.sku': {'$in': ['2baf524e08e3c9daa7caab5f']}}                                                      
```

Old code, default coercion:
```
In [10]: %timeit mongo._mongotize(source, resource)                                                                                                          
9.12 µs ± 18.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

Old code, no coercion:
```
In [13]: eimm.config.DOMAIN["contacts"]["query_objectid_as_string"] = True            
                                                                       
In [14]: %timeit mongo._mongotize(source, resource)                                                                                                          
7.74 µs ± 32.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

New code no coercion (restarted ipython):
```
In [9]: mongo._mongotize(source, resource)                                                                                                                   
Out[9]: 
{'_id': ObjectId('5e20d925927678c6959b3d36'),
 'rows.sku': {'$in': [ObjectId('2baf524e08e3c9daa7caab5f')]}}

In [10]: %timeit mongo._mongotize(source, resource)                                                                                                          
8.93 µs ± 21.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

New code with less coercion
``` 
In [12]: eimm.config.DOMAIN["contacts"]["query_objectid_as_string"] = True                                                                                   

In [13]: %timeit mongo._mongotize(source, resource)                                                                                                          
7.53 µs ± 79.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```