@ehiggs 2020-01-17
Oops, I wasn't actually calling `get_schema_type` 😞 
Added a commit that fixed this and also fixed a test that no longer demonstrated what it was supposed to (`test_get_lookup_field_as_string`, and I think it wasn't correct anyway since it didn't set the config variable it was claiming to be testing).