@ehiggs 2020-01-17
Trying to reproduce the travis error:

```
➜  eve git:(always-coerce-id-fields) ✗ .tox/py27/bin/py.test -k TestGet
==================================================================== test session starts ====================================================================
platform darwin -- Python 2.7.16, pytest-4.6.9, py-1.8.1, pluggy-0.13.1
rootdir: /Users/ehiggs/src/eve, inifile: pytest.ini, testpaths: eve/tests
collected 803 items / 702 deselected / 101 selected                                                                                                         

eve/tests/methods/get.py .....................................................................................................

======================================================== 101 passed, 702 deselected in 32.48 seconds ========================================================
➜  eve git:(always-coerce-id-fields) ✗ git push ehiggs
Everything up-to-date
➜  eve git:(always-coerce-id-fields) ✗ .tox/py27/bin/py.test -k TestGet
==================================================================== test session starts ====================================================================
platform darwin -- Python 2.7.16, pytest-4.6.9, py-1.8.1, pluggy-0.13.1
rootdir: /Users/ehiggs/src/eve, inifile: pytest.ini, testpaths: eve/tests
collected 803 items / 702 deselected / 101 selected                                                                                                         

eve/tests/methods/get.py .....................................................................................................

======================================================== 101 passed, 702 deselected in 32.56 seconds ========================================================
➜  eve git:(always-coerce-id-fields) ✗ .tox/py27/bin/py.test -k TestGet
==================================================================== test session starts ====================================================================
platform darwin -- Python 2.7.16, pytest-4.6.9, py-1.8.1, pluggy-0.13.1
rootdir: /Users/ehiggs/src/eve, inifile: pytest.ini, testpaths: eve/tests
collected 803 items / 702 deselected / 101 selected                                                                                                         

eve/tests/methods/get.py ........................................................................F............................

========================================================================= FAILURES ==========================================================================
______________________________________________ TestGet.test_get_where_mongo_objectid_as_string_but_field_is_id ______________________________________________

self = <eve.tests.methods.get.TestGet testMethod=test_get_where_mongo_objectid_as_string_but_field_is_id>

    def test_get_where_mongo_objectid_as_string_but_field_is_id(self):
        skus = self.to_list_string([item["sku"] for item in self.item_rows])
        where_in = '{"tid": "%s", "rows.sku": { "$in": %s} }' % (self.item_tid, skus)
        response, status = self.get(self.known_resource, "?where=%s" % where_in)
        self.assert200(status)
        resource = response["_items"]
        self.assertEqual(len(resource), 0)
    
        self.app.config["DOMAIN"]["contacts"]["query_objectid_as_string"] = True
        response, status = self.get(self.known_resource, "?where=%s" % where_in)
        self.assert200(status)
        resource = response["_items"]
>       self.assertEqual(len(resource), 1)
E       AssertionError: 0 != 1

eve/tests/methods/get.py:255: AssertionError
================================================================== short test summary info ==================================================================
FAILED eve/tests/methods/get.py::TestGet::test_get_where_mongo_objectid_as_string_but_field_is_id - AssertionError: 0 != 1
=================================================== 1 failed, 100 passed, 702 deselected in 32.01 seconds ===================================================

```

feelsbadman.png