@ehiggs-deliverect 2020-01-20
Found the issue with possibly failing tests. The tests were generating 0 - 5 rows but we want at least one row to use the sku in there. This is fixed in the latest patch.