@ehiggs 2020-01-20
I found that the tests were inconsistent because there were 0 - 5 rows generated. The new test requires 1 row with an sku to be generated so I updated the row generation to always make at least one.