@ehiggs 2020-01-21
Fixed a bug where a get request like `where?{"tid": { "$in": ["abc..."]}}` would fail because in the recursion "$in" isn't in the schema and would prevent the [...] from being converted to ObjectIds.